Migrate login endpoint, tests fail
