+++ server/service/handler_test.go
@@ -16,46 +16,9 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/throttled/throttled/v2/store/memstore"
 )
-
-func TestAPIRoutes(t *testing.T) {
-	ds := new(mock.Store)
-
-	svc := newTestService(ds, nil, nil)
-
-	r := mux.NewRouter()
-	limitStore, _ := memstore.New(0)
-	ke := MakeFleetServerEndpoints(svc, "", limitStore, kitlog.NewNopLogger())
-	kh := makeKitHandlers(ke, nil)
-	attachFleetAPIRoutes(r, kh)
-	handler := mux.NewRouter()
-	handler.PathPrefix("/").Handler(r)
-
-	routes := []struct {
-		verb string
-		uri  string
-	}{
-		{
-			verb: "POST",
-			uri:  "/api/v1/fleet/login",
-		},
-	}
-
-	for _, route := range routes {
-		t.Run(fmt.Sprintf(": %v", route.uri), func(st *testing.T) {
-			recorder := httptest.NewRecorder()
-			handler.ServeHTTP(
-				recorder,
-				httptest.NewRequest(route.verb, route.uri, nil),
-			)
-			assert.NotEqual(st, 404, recorder.Code)
-			assert.NotEqual(st, 405, recorder.Code, route.verb) // if it matches a path but with wrong verb
-		})
-	}
-}
 
 func TestAPIRoutesConflicts(t *testing.T) {
 	ds := new(mock.Store)
